Clustered BLAST now uses BLAST+ and lastal, and, move most of the code to clusterBLAST.pm

+++ lib/clusterBLASTp.pm
@@ -3,11 +3,15 @@
 package clusterBLASTp;
 require Exporter;
 use strict;
+use DBI;
 our (@ISA,@EXPORT);
 @ISA = qw(Exporter);
 @EXPORT = qw{cluster parseClustering expandByClusters clusteringIntoDb
              proteinsToFaa formatBLASTp runBLASTp proteinDbSize
-             parseBLASTpHits saveBLASTpHits};
+             parseBLASTpHits saveBLASTpHits
+             clusteredBLASTp
+             removeDuplicates
+          };
 
 # Writes the reduced fasta to out and the clustering to out.clstr
 # Saves the cd-hit log in out.log
@@ -91,28 +95,61 @@ sub clusteringIntoDb($$) {
   unlink($tmpFile);
 }
 
-sub formatBLASTp($$) {
-  my ($formatdb, $faa) = @_;
-  die "No such executable: $formatdb\n" unless -x $formatdb;
+# The directory should contain makeblastdbd, if $blastPlus, or formatdb otherwise.
+# Uf $blastPlus, puts the database in $db.plusdb
+sub formatBLASTp($$$) {
+  my ($exeDir, $faa, $blastPlus) = @_;
+  die "Not a directory: $exeDir\n" unless -d $exeDir;
   die "No such file: $faa\n" unless -e $faa;
-  my $formatCmd = "$formatdb -p T -i $faa";
-  system($formatCmd) == 0 || die "formatdb failed:\n$formatCmd\n$!";
+  my $formatCmd;
+  if ($blastPlus) {
+    die "No such executable: $exeDir/makeblastdb\n" unless -x "$exeDir/makeblastdb";
+    $formatCmd = "$exeDir/makeblastdb -in $faa -dbtype prot -out $faa.plusdb -logfile /dev/null";
+  } else {
+    die "No such executable: $exeDir/formatdb\n" unless -x "$exeDir/formatdb";
+    $formatCmd = "$exeDir/formatdb -i $faa -p T";
+  }
+  system($formatCmd) == 0 || die "makeblastdb failed:\n$formatCmd\n$!";
 }
 
-# The input parameters must include blastall (the path tot he
-# executable), query (a sequence), and db (the fasta database) Returns
-# a reference to list of hashes, each with subject, identity (as a
+# Required members of the input hash:
+# exeDir (containing either original BLAST blastall or BLAST+'s blastp),
+# query (a sequence),
+# db (the fasta database).
+# Optional arguments:
+#  blastPlus -- if not specifiedif db.plusdb.pin exists, then
+#    it uses blast+, otherwise it uses blastall.
+#  eValue -- default 1e-3
+#  nCPUs -- default 8
+#  dbSize -- default 0 (the actual size)
+#  maxHits -- defaults to 10,000
+#
+# Returns a reference to list of hashes, each with subject, identity (as a
 # fraction, not a percentage), alnLength, nMismatch, nGapOpens,
 # qBegin, qEnd, sBegin, sEnd, eValue, bits
 sub runBLASTp {
   my (%param) = @_;
-  my $blastall = $param{blastall} || die "Must specify blastall";
-  die "No such executable: $blastall\n" unless -x $blastall;
+  my $exeDir = $param{exeDir} || die "Must specify exeDir";
+  die "Not a directory: $exeDir" unless -d $exeDir;
   my $query = $param{query} || die "Must specify query";
   die "Invalid query sequence" unless $query =~ m/^[A-Z]+$/;
   my $db = $param{db};
   die "Must specify db" if !defined $db || $db eq "";
-  die "No such file: $db.pin" unless -e "$db.pin";
+  my $maxHits = $param{maxHits};
+  $maxHits = 10000 if !defined $maxHits;
+  die "Invalid maxHits" unless $maxHits =~ m/^\d+$/;
+
+  my $blastPlus = $param{blastPlus};
+  $blastPlus = -e "$db.plusdb.pin" if !defined $blastPlus;
+
+  if ($blastPlus) {
+    die "No such executable: $exeDir/blastp" unless -x "$exeDir/blastp";
+    die "No such file: $db.plusdb.pin" unless -e "$db.plusdb.pin";
+  } else {
+    die "No such executable: $exeDir/blastall" unless -x "$exeDir/blastall";
+    die "No such file: $db.pin" unless -e "$db.pin";
+  }
+
   my $minEValue = $param{eValue} || 1e-3;
   my $nCPUs = $param{nCPUs} || 8;
   my $dbSize = $param{dbSize} || 0;
@@ -124,9 +161,18 @@ sub runBLASTp {
   close($fhFaa) || die "Error writing to $tmpFaa";
 
   my $tmpHits = "$tmpPre.hits";
-  my $blastCmd = qq{$blastall -p blastp -e $minEValue -z $dbSize -a $nCPUs -i $tmpFaa -d $db -F "m S" -m 8}
-    . " -b 10000 -v 10000 -o $tmpHits >& /dev/null";
-  system($blastCmd) == 0 || die "blastall failed:\n$blastCmd\n$!";
+  my $blastCmd;
+  if ($blastPlus) {
+    $blastCmd = "$exeDir/blastp -evalue $minEValue -num_threads $nCPUs -query $tmpFaa -db $db.plusdb"
+      . " -max_target_seqs $maxHits -outfmt 6 -out $tmpHits";
+    $blastCmd .= " -dbsize $dbSize" if $dbSize;
+    $blastCmd .= " >& /dev/null";
+  } else {
+    $blastCmd = qq{$exeDir/blastall -p blastp -e $minEValue -z $dbSize -a $nCPUs -i $tmpFaa -d $db -F "m S" -m 8}
+      . " -b $maxHits -v $maxHits -o $tmpHits >& /dev/null";
+  }
+  die "Run: $blastCmd\n" if ! $blastPlus;
+  system($blastCmd) == 0 || die "runBLASTp() failed:\n$blastCmd\n$!";
   unlink($tmpFaa);
 
   my $hits = parseBLASTpHits($tmpHits);
@@ -134,12 +180,15 @@ sub runBLASTp {
   return($hits);
 }
 
+# Returns a reference to a list of rows, each as a hash
+# This also works on last BLAST-like format, by skipping lines that start with #
 sub parseBLASTpHits($) {
   my ($file) = @_;
   my @out = ();
   open(my $fhHits, "<", $file) || die "Cannot read $file";
   while (my $line = <$fhHits>) {
     chomp $line;
+    next if $line eq "#" || $line =~ m/^# /;
     my (undef, $subject, $identity, $alen, $mm, $gap, $qBegin, $qEnd, $sBegin, $sEnd, $eValue, $bits)
       = split /\t/, $line;
     die "Cannot parse\n$line\nfrom blastp" unless defined $bits && $bits =~ m/\d/;
@@ -210,6 +259,98 @@ sub proteinsToFaa($$$) {
     print $fh ">" . $proteinId . "\n" . $seq . "\n";
   }
   close($fh) || die "Error writing to $faaOut";
+}
+
+# The input hash must include the arguments
+# query (a sequence),
+# clusterDb (the BLAST+ database file name, without the .plusdb suffix),
+# dbh (the database handle, for expanding clusters and fetching protein sequences),
+# nCPUs,
+# bin (the directory with lastdb, lastall, and subdirectories blast+ or blast),
+# maxHits (a list with the maximum #hits for the first round and for the second round)
+#
+# Optional argument:
+# scale -- size of full database / size of clustered database (defaults to 1, only
+#   appropriate if you do not really care about the evalues)
+# eValue -- default 1e-3
+# quiet -- if not set, outputs HTML comments about what it is doing.
+#
+# Returns a list of hits, as from parseBLASTpHits
+sub clusteredBLASTp {
+  my (%in) = @_;
+  my $query = $in{query} || die "Must specify query";
+  die "Invalid query" unless $query =~ m/^[A-Z]+$/;
+  my $clusterDb = $in{clusterDb} || die "Must specify clusterDb";
+  my $dbh = $in{dbh} || die "Must specify dbh";
+  my $nCPUs = $in{nCPUs} || die "Must specify nCPUs > 0";
+  my $bin = $in{bin} || die "Must specify bin";
+  my $maxHits = $in{maxHits} || die "Must specify maxHits";
+  my ($nMaxHits1, $nMaxHits2) = @$maxHits;
+  die "Invalid maxHits" unless $nMaxHits1 > 0 && $nMaxHits2 > 0;
+  my $scale = $in{scale} || 1;
+  my $eValue = $in{eValue} || 1e-3;
+  my $quiet = $in{quiet} || 0;
+
+  print "<P><small>Running BLASTp.</small>\n" unless $quiet;
+  my $hits1 = runBLASTp('exeDir' => "$bin/blast+",
+                        'blastPlus' => 1,
+                        'query' => $query, 'db' => $clusterDb,
+                        'eValue' => $eValue, 'maxHits' => $nMaxHits1, 'nCPUs' => $nCPUs);
+  return [] if @$hits1 == 0;
+  my @clusterIds = removeDuplicates(map $_->{subject}, @$hits1);
+  splice @clusterIds, $nMaxHits1;
+  print "<small>Processing " . scalar(@clusterIds) . " clusters and re-running BLASTp.</small>\n"
+    unless $quiet;
+  my $proteinIds = expandByClusters(\@clusterIds, $dbh);
+  splice @$proteinIds, $nMaxHits2;
+  print "<!-- expanded to " . scalar(@$proteinIds) . " -->\n" unless $quiet;
+  my $tmpPre = ($ENV{TMPDIR} || "/tmp") . "/clusterBLASTp.$$";
+  proteinsToFaa($proteinIds, "$tmpPre.faa", $dbh);
+  print "<!-- fetched protein sequences -->\n" unless $quiet;
+
+  my $inFile = "$tmpPre.in";
+  open(my $fh, ">", $inFile) || die "Cannot write to $inFile";
+  print $fh ">query\n$query\n";
+  close($fh) || die "Error writing $inFile";
+
+  my $lastal = "$bin/lastal";
+  my $lastdb = "$bin/lastdb";
+  foreach my $x ($lastal, $lastdb) {
+    die "No such executable: $x\n" unless -x $x;
+  }
+  my @lastCmds = ("$lastdb -P $nCPUs -p $tmpPre.lastdb $tmpPre.faa",
+                  "$lastal -m $nMaxHits2 -f BlastTab -P $nCPUs $tmpPre.lastdb $inFile > $tmpPre.hits");
+  foreach my $cmd (@lastCmds) {
+    system($cmd) == 0 || die "last failed:\n$cmd\n$!";
+  }
+  my $hits = parseBLASTpHits("$tmpPre.hits");
+  unlink("$tmpPre.faa");
+  unlink("$inFile");
+  unlink("$tmpPre.hits");
+  foreach my $suffix (qw{bck des prj sds ssp suf tis}) {
+    unlink("$tmpPre.lastdb.$suffix");
+  }
+  # Limit the number of hits
+  splice @$hits, $nMaxHits2;
+  # Scale the evalues and filter by evalue
+  foreach my $hit (@$hits) {
+    $hit->{eValue} = sprintf("%.0g", $scale * $hit->{eValue});
+  }
+  my @hits2 = grep $_->{eValue} <= $eValue, @$hits;
+  return \@hits2;
+}
+
+# Remove duplicates from a list, and return the new list
+sub removeDuplicates {
+  my (@in) = @_;
+  my %seen = ();
+  my @out = ();
+  foreach my $value (@in) {
+    next if exists $seen{$value};
+    $seen{$value} = 1;
+    push @out, $value;
+  }
+  return @out;
 }
 
 1;
